+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on Keep a Changelog (https://keepachangelog.com), and this p
 - Add missing unit test for `Sgn` in `EvalExpr`.
 - Add missing unit test for `Int` in `EvalExpr`.
 - Fix error in `Int` when negative value is already an int.
+- Add missing unit test for `Sqr` in `EvalExpr`.
 
 ## [Unreleased] - 2026-02-14
 ### Changed